IT FIXED ITSELF
