+++ .breakpoints
@@ -1,12 +1,3 @@
 {
-  "files": {
-    "program.ev": [
-      {
-        "id": "2fa555a3-b81b-4036-a46f-7728bfb32635",
-        "line": 1,
-        "version": 1101,
-        "index": 0
-      }
-    ]
-  }
+  "files": {}
 }
